pyproject.toml
[project]
name = "llm-complete-command"
version = "2.1.1"
description = "Use LLM to generate commands for your shell"
readme = "README.md"
authors = [{name = "Ryan Patterson"},{name = "Scott Hyndman"}]
license = {text = "Apache-2.0"}
classifiers = ["License :: OSI Approved :: Apache Software License"]
dependencies = [
    "better-exceptions>=0.3.3",
    "llm",
    "loguru>=0.7.3",
    "platformdirs>=4.5.1",
    "prompt_toolkit>=3.0.43",
    "pyyaml>=6.0.3",
    "yaspin>=3.4.0",
]

[project.urls]
Homepage = "https://github.com/shyndman/llm-complete-command"
Changelog = "https://github.com/shyndman/llm-complete-command/releases"
Issues = "https://github.com/shyndman/llm-complete-command/issues"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.entry-points.llm]
complete_command = "llm_complete_command"

[tool.pyright]
include = ["."]
ignore = [".venv"]

[dependency-groups]
dev = [
    "basedpyright>=1.37.4",
]
